requirements.txt
ccxt
pandas
python-dotenv
pyyaml
streamlit
tenacity
pytest
